Split tag and epoch stats
The goal of this commit is to align Dramsim3 stats logging with vanilla_stats in
the manycore

This commit creates two sets of Counters (Epoch and Tag) and associated Printing
Functions (PrintEpochStats and PrintTagStats) within each class
memory_system->dram_system->controller->simple_stats

Within each set there are three types:
{set}_counters_ {set}_vec_counters {set}_histo_counts_

The only difference between the two sets of Counters is the period when they
print to a file and reset their counters

Epoch - prints every "epoch_period" cycle which is set in each dramsim3 config
file after it prints a period, the stats are accumulated in the total counter
and then reset to 0

Tag - prints whenever "print_stat_v_i" in manycore_tb_top is set high, along
with the associated tag in "print_stat_tag_i". This is the same with how
vanilla_stats.csv gets logged. After it prints the counter is reset to 0, but
they are not accumulated.