fix the infinity loop bug encountered in the R->W dependency
Fix the BUG of infinity loop encountered in the R->W dependency. It should be handled prior than write. When the R->W dependency occurs, we should schedule read queue for the next cycle, which aims to serve the corresponding read request before the write, and the variable ' write_draining' should not be overwritten by the write buffer size.